my_string

+++ lib/my/my_string/my_strlen.c
@@ -5,7 +5,9 @@
 ** display number character
 */
 
-int my_strlen(char const *str)
+#include <stddef.h>
+
+size_t my_strlen(char const *str)
 {
     int i = 0;
